feat: add unified jobs page
